concatdec: change data type to suppress warning about limited range of data type on some platforms.

+++ libavformat/concatdec.c
@@ -68,7 +68,7 @@ static int add_file(AVFormatContext *avf, char *filename, ConcatFile **rfile,
     av_free(filename);
 
     if (cat->nb_files >= *nb_files_alloc) {
-        unsigned n = FFMAX(*nb_files_alloc * 2, 16);
+        size_t n = FFMAX(*nb_files_alloc * 2, 16);
         ConcatFile *new_files;
         if (n <= cat->nb_files || n > SIZE_MAX / sizeof(*cat->files) ||
             !(new_files = av_realloc(cat->files, n * sizeof(*cat->files))))
